third_party/blink/web_tests/external/wpt/svg/animations/svglengthlist-animation-unitType-expected.txt

This is a testharness.js-based test.
[FAIL] Test change of unit type for SVGLengthList animation.
  assert_equals: expected 5 but got 2
Harness: the test ran to completion.
